+++ action_server/tests/action_server_tests/test_cli.py
@@ -43,22 +43,22 @@ def test_help(str_regression):
 def test_migrate(database_v0):
     from action_server_tests.fixtures import robocorp_action_server_run
 
-    from sema4ai.action_server.migrations import db_migration_pending
+    from sema4ai.action_server.migrations import MigrationStatus, db_migration_status
 
     db_path = database_v0
-    if db_migration_pending(db_path):
-        robocorp_action_server_run(
-            [
-                "migrate",
-                "--datadir",
-                str(db_path.parent),
-                "--db-file",
-                str(db_path.name),
-            ],
-            returncode=0,
-        )
+    assert db_migration_status(db_path) == MigrationStatus.NEEDS_MIGRATION
+    robocorp_action_server_run(
+        [
+            "migrate",
+            "--datadir",
+            str(db_path.parent),
+            "--db-file",
+            str(db_path.name),
+        ],
+        returncode=0,
+    )
 
-        assert not db_migration_pending(db_path)
+    assert db_migration_status(db_path) == MigrationStatus.UP_TO_DATE
 
 
 def test_default_datadir(tmpdir):